[FE] MOVE: 폴더 명 변경 및 에러 처리 #58
