+++ src/frontend/package.json
@@ -15,6 +15,7 @@
     "@sentry/browser": "^8.48.0",
     "@sentry/react": "^8.48.0",
     "@tanstack/react-query": "^5.64.1",
+    "@types/youtube": "^0.1.0",
     "axios": "^1.7.9",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
