+++ src/frontend/src/components/Sidebar/MemberList/index.css.ts
