+++ src/frontend/src/components/Sidebar/Playlist/index.css.ts
