+++ src/frontend/src/components/Sidebar/index.css.ts
