+++ src/frontend/src/components/Sidebar/index.tsx
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
-import { Chat } from './chat';
-import { MemberList } from './memberList';
-import { Playlist } from './playlist';
+import { Chat } from './Chat';
+import { MemberList } from './MemberList';
+import { Playlist } from './Playlist';
 import SidebarChat from '@/assets/img/SidebarChat.svg';
 import SidebarPlaylist from '@/assets/img/SidebarPlaylist.svg';
 import SidebarVoicechat from '@/assets/img/SidebarVoicechat.svg';
 import SidebarMember from '@/assets/img/SidebarMember.svg';
 import Add from '@/assets/img/Add.svg';
-import { CommonInput } from '@/components/common/input';
-import { CommonButton } from '@/components/common/button';
+import { CommonInput } from '@/components/common/Input';
+import { CommonButton } from '@/components/common/Button';
 import { chatDataTest } from '@/assets/data/chatDataTest';
 import { Wrapper, Nav, Content, NavButton, InputContainer, ChatContainer } from './index.css';
 import { SidebarType } from '@/types/enums/SidebarType';
