+++ src/frontend/src/components/common/Button/index.css.ts
@@ -25,7 +25,7 @@ export const StyledButton = styled.button<{
   font-size: 14px;
   border: ${({ border }) => border};
   border-radius: ${({ borderradius }) => borderradius};
-  justify-content:${({ justifycontent }) => justifycontent};
+  justify-content: ${({ justifycontent }) => justifycontent};
   cursor: pointer;
   transition:
     background-color 0.3s,
@@ -37,7 +37,7 @@ export const StyledButton = styled.button<{
     opacity: ${({ color }) =>
       color === ButtonColor.TRANSPARENT ? 0.8 : 1}; /* 투명일 때 hover 효과 */
   }
-\
+
   ${({ disabled }) =>
     disabled &&
     css`
