+++ src/frontend/src/components/common/SmallProfile/index.css.ts
