+++ src/frontend/src/pages/RegisterPage/index.tsx
@@ -1,7 +1,7 @@
 import { useRef, useState } from 'react';
 import { LogoButton } from '@/components/common/LogoButton';
 import { ButtonColor } from '@/types/enums/ButtonColor';
-import { CommonButton } from '@/components/common/button';
+import { CommonButton } from '@/components/common/Button';
 import {
   CommonInput,
   CommonLabel,
